requirements.txt
python-dotenv==1.0.1
requests==2.32.3
tenacity==8.4.2
pydantic==2.9.2
pandas==2.2.2
tqdm==4.66.5
